requirements.txt
streamlit
torch
torchvision
pillow
pillow-heif

fastapi
uvicorn

python-multipart